+++ functions/mainFunctions/deleteUser.js
