+++ functions/mainFunctions/newUser.js
