+++ functions/utils/refsToDocs.js
@@ -1,0 +1,7 @@
+module.exports = async (refArray) => {
+    return Promise.all(
+        refArray.map(ref => {
+            return ref.get();
+        })
+    );
+}
